Initial changes that allow ModelEvaluator to build.
Changed many instances of Peridigm to PeridigmNS.

Started work on a new constructor for ModelEvaluator.  A number of
objects will be passed in via RCP as opposed to being constructed
inside ModelEvaluator.

Currently ModelEvaluator is dependent on classes that will
probably be removed from the code base as the refactor progresses
(for example, VerletSolver).  For now, all these dependencies
are still there.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@423 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 2817fa8b067b783f35804f00d335720256818c9f