More work on refactor.
Things are now set up so that the Peridigm object
constructs and makes calls to the ModelEvaluator.
The materials list is passed to the ModelEvaluator
constructor, and the evalModel() function now takes
a workset (as opposed to inArgs and outArgs).  This
makes the code much cleaner and more flexible from
the peridynamics point of view.  The downside is
that our version of ModelEvaluator is
less of an abstraction of a general model than
the original Agile Components ModelEvaluator is. 



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@425 d9297710-12f9-482b-b11f-fb677dc88bbb