Stripped PeridigmNS::ModelEvaluator of things not related to the actual evaluation of models.  It has basically been reduced to a wrapper around a field manager.
There is quite a bit of commented-out code in Peridigm_ModelEvaluator.cpp,
the plan is to remove it once Peridigm is up and running.

The Verlet integrator and observer, as well as all the output code, is
currently not being compiled.  It was written to interface with ModelEvaluator,
and the accessor functions in ModelEvaluator have been removed.  I'm planning
to update the output code so that it can be called more directly.  This will
allow us to get Peridigm up and running quickly and get the tests passing.
Then we can refactor the output code as we've discussed.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@427 d9297710-12f9-482b-b11f-fb677dc88bbb