Incremental progress on Peridigm refactor.
The following appear to work:
- Instantiation of materials
- Creation of discretization
- Application of initial velocities
- Creation of workset
- Creation of ModelEvaluator
- Initialization of materials (distinct from their instantiation)

Peridigm::execute() now makes a single call to
modelEvaluator->evalModel(workset) and the results
appear sane.

The next step is to get the integrator up and
running.  The integration loop will be in the
Peridigm::execute() function, where we at least
know we can call ModelEvaluator::evalModel() to
get the accelerations.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@428 d9297710-12f9-482b-b11f-fb677dc88bbb