Incremental checkin towards cleaning up the build scripts.  More to come.
This modifies the main CMakeLists.txt file by using 'include_directories' and 'link_directories.' 
Nexted folders do not have to include that command.  Libraries are seperated from their path.  This 
is useful for at least two reasons:
1) During link its much easier to read the link line since the libraries are included without their path
2) Paths to link directories can be created at the highest level and then passed in to nested 
directories -- only one file has to be modified.  A simplication (less code) in the nested CMakeList.txt files.

I also took a shot at creating a python script for symbolic links.  Its a rather simple 
version but it works on a couple of directories that have the simple case of file only links.  If we like 
this approach I can enhance the script to include the nested folders variety that Dave has implemented 
in the test/verification directory as a cmake script.

Deleted a very old directory that I created a long time ago -- we don't use any of this code.
 


git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@432 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: a45ec94983a4f04faeb052639f08fe927e39ecb3