Progress toward rebalance.
Peridigm now has a rebalance function that creates
a new discretization based on a rebalanced PdGridData
object.

The next task is to create new vectors and maps, and the
import/export objects that will allow the data from the
old data containers to be transfered to the new data containers.

The first attempt at rebalance will focus on a test case with
the elastic material model and no bond breaking to minimize
the data that must be redistributed.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@440 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 5059f29a0aa2db3545cc72cb33e5eb8a36c174e2