Fix vtkIO to more properly handle types.
Mike: I think this fixes the issue you had with putting out 
and integer.  I added an integer to the unit test demo -- it runs but 
I'm at home some I can't really run paraview to see if it works or not. 
I think its fine.  

Solution turned out to be very cool.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@443 d9297710-12f9-482b-b11f-fb677dc88bbb