Progress on vtkDiff.py
Now works with the "uniform tolerance" option.  In this case
the two data sets must have the same named point data fields,
and all the fields are compared with the specified tolerance.

Appears to work fine if the number of processors is different
for the two different data sets.  This is important for the
automated tests, where we want to archive only the serial results
and then run the test in both serial and parallel.

The next task is to get the ".comp" files hooked up like they
were before.  This will allow the user to specify the fields
to be compared and the tolerance for each field.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@455 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 0d29bd8055ac5b7d0f075b3417e9d6931665e345