Adds verification test for elastic-plastic damage model.
Dave,

Basically -- this is a copy of an existing test for the elastic case.  The input is set up 
for the elastic-plastic model but with an extra command in the xml: "Test".  This causes 
the plasticity model to set the yield stress to the largest possible value on the machine:
	/*
	 * Set the yield stress to a very large value: this in effect makes the model run elastic -- useful for testing
	 */
	if(params.isType<string>("Test"))
		m_yieldStress = std::numeric_limits<double>::max( );

Under these conditions, the material will not yield.  I don't really know how to hook up the diffing 
aspect for this test but it does run through to completion.  Just now sure if it 
is giving identical answers to the elastic case (as it should).



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@477 d9297710-12f9-482b-b11f-fb677dc88bbb