Starting working on the handling of constitutive data in Peridigm.
Goals:

1) Create a data manager class to handle the storage
and accessing of constitutive data.

2) The data manager will give both the material models
and the I/O system access to variables at either STATE_N,
STATE_NP1, or STATE_NONE.  We will no longer pass separate
Epetra_Multivectors to the material models; we will pass
a data manager instead and the Epetra_Vectors of constitutive
data will be accessed through it.

3) Treat states (e.g. STATE_N, STATE_NP1) as
entities within data manager.  For example, the data
manager should be able to create an empty state, copy
a state, etc.

This checkin:

Created a function in Peridigm::Material called
VariableSpecs() that returns a RCP to a std::vector
of FieldSpecs.  The idea is that Peridigm will query
the materials for a list of their field specs and
will allocate memory accordingly.  The VariableSpecs()
function will eventually replace NumScalarConstitutiveVariables(),
etc., and all the data pertaining to variables will be
accessable through a single call to VariableSpecs().



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@484 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 609729772d8f04c374bcecd2736512595342c233