Set up a vector of FieldSpecs for both material models.
This was actually a bit tricky due to some C++/stl subtleties.
Turns out that if a class has a copy constructor that is declared
explict, objects of that type cannot be put in a stl::vector. I
know there are some restrictions on what can be put in a stl::vector,
for example a default constructor must be available, but I'm not
really clear on why the 'explicit' causes problems.

John, I edited Field.h and took the 'explicit' off the copy
constructor for FieldSpec.  Let me know if you see a better
way of handling this.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@485 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 1e60fa8a64abfbfd9cdf683b17e1f03bee02b19b