Continued work on data handling.
Created a State class and a DataManager class.

Very much still in the planning stage here,
neither is even hooked up to the build system.

The idea is that the DataManager is the magic object
that gets passed to the material models and I/O system.
There will be an interface that provides access to the
underlying data for a given FieldSpec and step (N,
NP1, or NONE).

Inside the data manager, data will be organized within
State objects.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@486 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 5ffd71fb9a70f3fe1cda54efbb4173eb30cc5a26