Added stateArch member variable to FieldSpec.
Its type is:

enum FieldStateArchitecture {STATELESS=0, STATEFUL=1};

I updated the FieldSpec constructors.  There is a new one
that takes a FieldStateArchitecture argument.  The old one
is still there and assigned STATELESS to the stateArch.

Updated all the hard-coded FieldSpecs to include a stateArch.

Also updated the equality and inequality operators for FieldSpec.

Note to John, I did not update FieldSpec::operator< because
I wasn't sure how you'd want it.  So, it currently does nothing
with stateArch.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@499 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 774b1307b5b2a70f8ce45aaa434cb3d8d625ff14