Yet another checkin on DataManager work.
The DataManager is now fleshed out to the point where
the individual states are created (NONE, N, and NP1)
and the data can be accessed via a function:

DataManager::getData(Field_NS::FieldSpec fieldSpec, Field_NS::FieldSpec::FieldStep fieldStep)

which returns a RCP to an Epetra_Vector.

For now, both the old data management system and the new
data management system are in the code.  The old system
is still hooked up.  After some additional testing I'll
flip the switch and Peridigm will use the new system.

Note that the memory footprint will be double what it
should be as long as both systems are in place.  It will
shrink down to what it's supposed to be when the original
system is removed.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@501 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 3bb6a1a26899e9a7ee13a1e4f7d2c966d4725236