Expanded DamageModel so that it now has the ability to register for data in the DataManager.
It is set up in the same way as Material.

Recall that each Material has a DamageModel (which
can be null).  At initialization, the Material queries
its DamageModel for a list of FieldSpecs corresponding
to the data the DamageModel needs to store.  This list
of FieldSpecs is then appended to the Material's list
of FieldSpecs, which is subsequently passed to the
DataManager so that space can be allocated.

See Peridigm_LinearElasticIsotropicMaterial.cpp and
Peridigm_CriticalStretchDamageModel.cpp for an example.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@506 d9297710-12f9-482b-b11f-fb677dc88bbb