This test is not turned on but the input syntax allows for it to run.
This test exercises the plasticity model and it appears to run correctly.

A small mesh is created.  Initial velocities of points elongate the cube.  
The cube continues to elongate until all of the initial energy is consumed -- 
this leaves the cube in a plastically deformed state.

There is a small amount of elastic energy left over -- this energy results 
in small motions/vibrations around the plastically deformed configuration.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@512 d9297710-12f9-482b-b11f-fb677dc88bbb