This fixes the Zoltan related test failures in Peridigm.
There was a semantics  change associated with Zoltan.  Prior to 
the recent zoltan changes, there was not a requirement for Peridigm 
to Zoltan_Initialize().  Zoltan_Initilize was simply calling MPI_Init 
which is something that we are already doing.  With the zoltan update,
Zoltan_Initialize() is doing additional work beyond calling MPI_Init.
I added Zoltan_Initialize() as recommended by Lee Ann and the test 
failures went away.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@527 d9297710-12f9-482b-b11f-fb677dc88bbb