Initial DAKOTA connection.
This introduces a new verification test where DAKOTA drives Peridigm. 
It's essentially the inverse of the Compression_2x1x1 test. In the compression
test, a material is specified, and initial conditions are set, then the test checks
the computed forces, velocities, etc.

In this test, forces are specified, and DAKOTA performs a calibration study, 
making many calls to Peridigm 
to determine the material properties that best fit the given data. 

This test requires a DAKOTA installation, and defaults to "OFF". If you want to 
enable the test, put a 
-D USE_DAKOTA
in your Peridigm build script.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@532 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 838d8318a36cf095cca6aeea375defcdcea4f368