More progress toward refactor to DataManager.
Removed dependence on DecompositionStates and then
removed the files related to DecompositionStates.

Removed the material accessors like NumScalarConstitutiveData()
and the code related to variable names, since all this is handled
with FieldSpecs in the DataManager now.

I think we're in position to tear out the remaining old-style
arguments to the material models and damage models.  That will
be the next task.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@536 d9297710-12f9-482b-b11f-fb677dc88bbb