Incremental checkin towards adding cracks to a mesh.
Basic idea/approach is as follows:
1) KdTree is first used to identify all bonds associated with a point
2) A filter is then used to identify bonds which should be removed in 
order to create a crack in the mesh

The filter has a particular virtual function.  The hope is that 
this form will allow for cracks, voids, and other features to be easily 
added.

This functionality needs a bit more work but not a lot.  Expect 
to finish this up sometime next week.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@547 d9297710-12f9-482b-b11f-fb677dc88bbb