Incremental checkin towards filtering bonds/neighborhoods in order to create cracks and other geometric features.
With this checkin, the default search algorithm/filter is 
completed and runs all of the unit tests except those 
that require 'self' in the neighborhood search.  That 
will be a different filter.

To Do:
* create filter that includes 'self' in the neighborhood search
* create a crack filter
* hook up the plumbing at the peridigm level first to 
handle the default case, and then to handle the self case, 
and finally to handle the 'crack'



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@548 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: d02bf8b4fa68b50f98d5983d3f8d67d667520faa