Incremental checkin for adding BondFilters to neighborhood search.
This completes the initial functionality.  Now all 
the unit tests run with the new filters in place.

There are two relatively simple filters:
1) neighborhood of x does not include x -- this is default
2) neigbhorhood of x includes 'x' -- this one must be 
explicitly constructed and passed in

To DO and up Next:
Build a filter that satifies the above interface and 
creates a crack.  Plan is to implement this 
by creating a plane and filter bonds that cut through 
the plane.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@550 d9297710-12f9-482b-b11f-fb677dc88bbb