Cleans up the CMakeList.txt file.
COMMENT:
My sense is that the current 'working' version of 
our CMakeLists.txt file is fragile wrt 'BOOST'.

The backup version that Dave sent me had the 
'VERSION' of boost hardwired.  Apparently that doesn't matter
because Dave was setting a 'BOOST_ROOT' to a different 
version in his shell script and that was working fine.

CHANGE:
I slightly changed the implementation of
linking with 'blas' and 'lapack' -- this version 
makes it much easier to use an optimized version 
if you wish.  Basically I set a BlasLapack_Libraries 
variable at the top.  Then the linker just looks 
to these variables.  This makes it easy to 
override these variable with your own 'blas/lapack' 
libs.

Hopefully, this will not break anyone's build.  I'm 
sure you will let me know:)  Of course, it builds on 
my machine.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@559 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: fee4b79527b874645896240bae9680dd389dc570