Yesterday, I was trying to run the expanding ring.  I made a mistake in the xml input deck for a fairly large mesh and it brought my machine down. We all agree that initial conditions in Peridigm are a 'hack'.  So I am proposing the following:
This is a sketelon outline of what an initial condition 
could look like in Peridigm.  Feedback is requested.

Basic Idea:
* Abstract InitialCondition class requires method "apply"
* Each new kind of initial condition simply overloads this method
* At the minimum, all initial conditions will be constructed using ParameterList peridigmParams
as an argument.
* Perhaps constructor should also take initialX, and pointers to 'velocity' and or something else 
that has access to all of the data arrays
* The 'InitialCondition' object is throw away.  We build it at the start, apply the 
initial conditions and then thow it away.

Example input for Q2Cylinder: take a look at src/initial_conditions/expanding_ring_IC.xml



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@561 d9297710-12f9-482b-b11f-fb677dc88bbb