Added a meaningful unit test for rebalance.
The test is an eight-point cube where two of
the points get swapped.  The rebalance should
then switch the owning processors for these
points and redistribute all the data associated
with them.

So far, the test asserts correct values for
global IDs and the mothership vectors.

More asserts will be added...



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@579 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: aaf9d15eb7dcc1dffbb9152ef2775039100db39e