This is an odd checkin.
This is an incremental checkin for adding a post 
time step calculation.  Its currently somewhat of 
a hack but I think this may prove to be an effective
way for doing user-unique post processing type calculations
for each time step.  I am currently working on the 
expanding_ring post processing for very large meshes.  
These meshes are perhaps too large to post-process 
with python scripts that pull the entire mesh into 
memory on one processor.  

I accidentally deleted src/Peridigm_OutputManager.hpp 
in my project with svn/eclipse so that's
why its here.  There are no changes to that.

There was a bug in src/Peridigm.cpp relating to 
the output manager and time.  I fixed that 
although the output manager does not currently 
output time?  We will need to fix that.  The 
correct time is currently available but if I add
it then there may be a bunch of diffs.  It only 
shows up in the PVD file -- maybe there won't be 
a bunch of diffs?

I added a new function in the vtkIO.py  -- helper
function to create a KdTree.

 



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@588 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 1e7f1bda1281a289f3da342ba086e8b31f437781