Expanded rebalance unit test (two processors) to include a check on bond damage.
The specific test is one in which the model is
an eight-point cube.  The current positions of
two of the nodes are swapped, and then rebalance
is called.  The resulting partitioning should
have these two nodes switched.

For the bond damage test, the bond damage values
are set to 1.0 (broken) for several bonds which
belong to the nodes that will be swapped.  Then,
after rebalance, the bond data are checked to
make sure that all bonds are intact except for
those that were broken prior to rebalance.  The
data indicating the broken bonds should travel
to the new processors along with all other data
associated with the swapped nodes.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@591 d9297710-12f9-482b-b11f-fb677dc88bbb