Post processing code for "expanding ring" demonstration.
Currently, there is some issue with time which I haven't 
worked out yet.  I am very close and the averaging quanties 
that I need to calculate.  With this code, I should be
able to scale up the problem size very large and not 
have to worry about post-processing after the fact.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@592 d9297710-12f9-482b-b11f-fb677dc88bbb