Closing in on rebalance.
The main part of this checkin is a new function
in DataManager called scatterToGhosts().  It
ensures that all processors with a copy of
a given point have the same data for that point.
It copies the value from the processor that owns
the point to all procesors with ghosted copies.
In order to get this to work, DataManager had to
be expanded to include a non-overlap map
corresponding to each overlap map.  This was
easy for the scalar and vector data, because
Peridigm already had these maps.  The non-overlapping
bond map must be created in DataManager.

The scatterToGhosts() function provides a fix
for what was essentially a logic error in
the rebalance of DataManager.  In order for
rebalance to work, all processors must have
the same value for overlapped points.  Otherwise,
Epetra will behave in an undefined way, since
it basically picks one of these values at random
and exports it to the rebalanced vector.  (The
concept of one processor 'owning' a given point
is a Peridigm idea, the overlap vectors do
not contain this logic; DataManager uses the
non-overlap maps to determine the owner.)

A small change to the output manager was also
required, since the vtk grid becomes invalid
after rebalance.  The output manager now checks
if a rebalance has occured and reinitializes
the grid if it has.

At this point all the tests will pass if you
turn on rebalance.  This can be done with a
hard-coded flag in Peridigm.  Setting this
flag forces a rebalance at every time step.
Enabling rebalance via the input deck is
forthcoming.

With rebalance turned on at every step, all of
the tests run smoothly except for the bond
breaking verification tests.  They still pass,
but the three-processor versions take about five
minutes to run, as opposed to the normal five
seconds.  Granted, they are rebalancing at every
time step, and there are 8000 time steps in these
tests.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@597 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: c441bf00541f8d98f2281ea10143d28d5683a8b3