Work toward handling the case where a point has zero neighbors.
Along the way I realized that a chunk of code
I wrote last week is totally unnecessary.  I
was calling scatterToGhosts() for bond data,
when in fact it needs to be called only for
scalar and vector data.  Bond data is stored
according to the neighborhood list format,
which is not overlapping.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@600 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 3c03dd034e6ee70d8e2c919fbf39e0b30fa4860b