Added a rebalance section to the input deck and created a rebalance regression test.
The rebalance section of the input deck is inside
the solver section and looks like this:

<ParameterList name="Rebalance">
  <Parameter name="Rebalance Frequency" type="int" value="100"/>
</ParameterList>

The new test consists of a 3x3 cube partitioned over
two processors.  The top face and bottom face
are given initial velocities such that they
move in opposite directions.  Rebalance is
called three times during the analysis.  The
resulting decompositions should be geometrically
partitioned.  All output quantities are checked
against the gold results, but the important thing
is the Proc_Num.  In theory, there is more than
one vaild solution to this problem, hence I declared
it a regression test as opposed to a verification test.
I.e., if there are code changes that cause this test
to fail, the code changes aren't necessarily wrong.

The paraview animation of this test is kinda neat,
it shows the evolving geometric decomposition if
you color by Proc_Num.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@602 d9297710-12f9-482b-b11f-fb677dc88bbb