Initial implementation of timer class.
The class has three main functions:

startTimer(string name)
stopTimer(string name)
printTimingData(std::ostream &out)

The timer class itself is a singleton.

To track the time spent in a given section of code,
use the startTimer and stopTimer functions:

PeridigmNS::Timer::self().startTimer("MyTimer");
...do some expensive calculations...
PeridigmNS::Timer::self().stopTimer("MyTimer");

The printTimingData function is called just prior
to Peridigm exiting.  It prints a table of the
various timing data.  I plan to extend this function
to print the min, max, and total time across the
different processors.  Currently it just prints the
times from the root processor, and the table is
pretty ugly.

I put a few timers in the code, they're all currently
in the integration loop.  This makes the code more
difficult to read, but we can improve on this by
breaking out a few functions from the integration
loop and moving the timer calls into these functions.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@621 d9297710-12f9-482b-b11f-fb677dc88bbb