Adds a parallel mesh reader using vtk. This does not build right now -- needs vtkMPI... and vtkParallel libs.
I am checking in vtkExodusReader.cxx -- this is from vtk-5.6.1.  I 
have commented out some annoying print statements that were somehow 
checked into the vtk repository.  When Dave and or Mike decides
to build the parallel pieces of vtk for our test platform,
they need to substitute this file for the one in the vtk release 5.6.1.

Please talk to me when you get around to doing that.  I have done this 
on my machine and it is fine.






git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@627 d9297710-12f9-482b-b11f-fb677dc88bbb