Several small 'typo' type fixes.
Incremental checkin towards 'crack opening' demonstration.

Main work relates to development of 'crack' opening 
calculation.  The demonstration code is 
basically set up.  However, I forgot/overlooked 
the fact that the jacobian also has to reflect 
the presence of the crack.  Need to do a small 
amount of work on that aspect for this demonstration 
code to run properly.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@629 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 2db1daf93e5edd2ea8d4de660886845a4b686e73