Add a function that computes a scale factor for the shear modulus in order to prevent the so-called 'dreaded surface effect' -- DSF.
After my analysis and brief discussion with Stewart,
I have concluded that for 'ordinary' isotropic 
materials, the DSF is purely a geometric artifact.

My current theory is that differences that now 
exist between the expanding ring calculations 
using Peridigm and Presto can be reconcilled 
with this scale factor.  

To DO: 
Need to figure out how to incorporate this into the 
plasticity model.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@638 d9297710-12f9-482b-b11f-fb677dc88bbb