Adds a new filter: crack plane filter that includes 'x' in the neighborhood(x).  Basically this means that a neighborhood of a point x includes the point x. Why?  Because this sort of neighborhood is required for calculating the jacobian.
Adds a unit test -- basically its one of the 
other unit tests but now the neighborhood list 
includes the point 'x'.

I am currently a bit baffled on the complexity 
involved:(  The filters are very nearly identical 
with and without 'self' but there are very subtle 
differences that are confusing.  I am still 
trying to develop the best way to handle this 
complexity.
 


git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@643 d9297710-12f9-482b-b11f-fb677dc88bbb