Progress toward finite-difference Jacobian.
A finite-difference Jacobian is now implemented
in the PeridigmNS::Material base class.  The
idea here is that this will be available for
any material regardless of whether or not
a consistent tangent is available.  If a
consistent tangent is available, the finite-
difference jacobian could be used for testing.

This checkin:

Fixed a couple of bugs in initial implementation
of finite-difference Jacobian.

Added a print function and a putScalar function
to DenseMatrix.  This class will be refactored
in general as the implicit solver matures.  Right
now it is simple but inefficient.

Extended the linear elastic material model unit
test to include a call to the Jacobian.  This test
needs to be further fleshed out.

IMPORTANT:

The finite-difference Jacobian works by calling
the internal force for individual points, one
at a time, in perturbed and unperturbed configurations.
The functions in PDMaterialUtilities make the assumption
that the first entries in the ownedIDs, X, Y,
weightedVolume, dilatation, and force vectors belong
to the ownedIDs.  The behavior of the finite-difference
Jacobian breaks this assumption.  Therefore, I implemented
new versions of computeDilataion and
computeInternalForceLinearElastic that take a list of
ownedIDs as an additional argument.  The Peridigm linear
elastic material model now calls the new versions of the
functions.  All the tests pass.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@661 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: d84f671353092e3b169d79ca068cc2cbfa7a9d8d