Finally beginning to make progress on the dsf verification calculations.
Lesson learned:
Weighted volume is extremely sensitive to the 
spatial distribution of points.  In fact it is 
so much so that to get a reasonable value 
as compared to an analytical value over a sphere
it is necessary to have a uniform mesh.  Mesh 
cannot be randomly generated like I had originally 
hoped.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@682 d9297710-12f9-482b-b11f-fb677dc88bbb