Under practical conditions, a value for the dsf must be calculated 3 times at each point and then the largest value is selected.  This is conceptually identical to Stewart's note for how he does this for 'bond' based materials.
The main reason for this is that at a given 
point, probing for XY, XZ, YZ shear will 
be mesh dependent and one or more of these 
probes could produce nonsense.  So, what is 
done is that we probe all 3 and then take the 
maximum value.  If for some reason all three 
are bad, then we just set the DSF=1.0 and it 
will have no effect.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@684 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: d5340afbc8732649110ca86674de706a1e4b3e63