This is part of a large refactoring of the 'io' directory.  The 'io' directory currently includes the following: * quick grid mesh generation * loadbalancing * parallel search and neighborhood creation * vtk output
The new structure is emerging.  Its not 
finalized but it will likely look like:
* mesh_input/quick_grid
* mesh_input/vtk
* neighborhood/
* utilities/
* load_bal

Over the next week or so, the number of tests 
in Peridigm through 'PdITI' will increase substantially 
because we will be carrying along both the deprecated 
and new configurations.  After I am convinced that the 
new configuration is running correctly we will need to 
refactor 'Peridigm' a small amount to use these new 
libraries.  Then we can eliminate the 'deprecated' 
source.





git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@690 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: be6e73bd634a473b355aab33777bf4d563026275