This is a bug fix related to the DSF. Thanks to Dave for helping me figure out what the issue was.
On my linux box, and apparently on all of our other 
build platforms, one of our unit tests was computing 
numerical zero as part of the dsf calculation. 
In this particular test, the deviatoric extension state is also 
zero and it is divided by the former numerical zero 
which has a cancellation effect that 
is perhaps physically correct except that on red-sky the numerics
do not work out so nicely.  I added a guard against this
by simply setting the DSF=1.0 when such pathological 
conditions occur. This has no effect since the deviatoric 
extension state is zero anyway.   Apparently, the deviatoric 
extension state tends to zero faster than its demoninator.
My observations are that the pathological 
conditions arise when there is only 1 bond at a point.  In this 
case, the deviatoric extension state is zero and the force state depends 
only on the bulk modulus.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@721 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 48614f4e169c7ec20f6b1f8dcd4133afb7517222