Fixes for quasi-statics in parallel.
The simple example now runs in both serial and parallel.

One confounding issue is that Eperta_CrsMatrix (and Epetra_FECrsMatrix)
is constructed with an Epetra_Map.  We went through a hassle a few
months ago to refactor Peridigm to use Epetra_BlockMap where possible.
The block maps have some nice features that make bookkeeping easier.
Unfortunately, filling an Epetra_FECrsMatrix with entries that are
arranged via an Epetra_BlockMap requires a bunch of index mapping,
since IDs in the block map do not correlate directly with IDs in
the non-block map.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@747 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 13f841988afffd7f598585f6c469ad667adbf058