Worked with John to modify the explicit time integrator so that the constitutive models get correct values for displacement_n and displacement_np1.  They were previously incorrect, but this caused no errors because the current constitutive models do not use this data.
Everything is consistent now except that the np1 velocity sent
to the constitutive models is actually n plus one half.  This
is, of course, how the explicit time integrator works, so there's
no error per se.  However, it is probably a good idea to make
this more transparent to constitutive model developers by creating
a vector in the DataManager called MIDSTEP_VELOCITY.  It would be
good as well to add functionality to the DataManager to allow fields
to be marked as invalid, for example to disallow access to velocity
at step np1 in the consitutive models (because this velocity is not
known at the time the constitutive models are called).

One other note, we discussed adding a "bootstrap" call to the
constitutive models prior to entering the explicit time integration
loop.  This would give us a more valid acceleration value for time
zero.  Currently, the acceleration at time zero is set to zero, and
thus the initial evaluation of the midstep velocity is not correct
(actually gives you the initial velocity).  This change would break
system-level tests that compare against hand calculations, since
the hand calculations assume the current time integration scheme.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@753 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 323840463087ed47f02879781cff325154cff09e