Added tests to utPeridigm_State unit test.
This unit test is set up so that it can be run
either in serial or in parallel with np = 2.
This is great for code consolidation, but
unfortunately the run_unit_test.py script
will create log files with the same names for
the two versions of the test, so whichever
one is run first will have its log file
overwritten.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@782 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: f0d79252bfad251f1a3e58023f4d0a100c28dfe5