Updated the build system to use FIND_PACKAGE(Trilinos).
This is cleaner than serching for each library individually,
which is what we had before.  We did this originally because
there was no FIND_PACKAGE for Trilinos at the time.

Also modified mirror.py so that it doesn't flood stdout
with a list of every symbolic link it removes.  It now
just reports the total number of symbolic links removed.

This work is preparation for using stk to read/write
exodus files.  Stk is a collection of libraries, so
searching for them individually would have been tedious.
Using FIND_PACKAGE(Trilinos) is better.

IMPORTANT:  At some point in the near future, Peridigm
will become dependent on netcdf.  Specifically, the
stk functionality for reading/writing Exodus files
depends on netcdf.  Netcdf is included in the trilinos
build of stk, not in Peridigm directly.  Initially, I'll
set up the build system such that netcdf is an option,
without which the Exodus functionality will be disabled.
This should avoid the pain of suddenly requiring a new
TPL.

Specifically, in the Trilinos build, the following
will be required:

-D TPL_ENABLE_Netcdf:BOOL=ON
-D Netcdf_INCLUDE_DIRS:PATH=<your_path_here>
-D Netcdf_LIBRARY_DIRS:PATH=<your_path_here>



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@786 d9297710-12f9-482b-b11f-fb677dc88bbb