Some further steps towards implicit time integration.
This adds a test to apply a nonzero Dirichlet boundary condition to the leftmost
particle of a two-particle system, and let the rightmost particle oscillate.
Time integration is implicit, with Newmark-Beta. 

Most of the initial desired functionality is now in place. Still need to clean up
the input deck a little, and address a map issue in parallel. 



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@827 d9297710-12f9-482b-b11f-fb677dc88bbb