Initial work on functions for copying data from one DataManager to another.
The use case is in the finite-difference Jacobian, where a miniature DataManager
is created for a single point and its neighbors.  Here, the miniature
DataManager is created with an Epetra_SerialComm, whereas the source DataManager
was created with an Epetra_MpiComm.  For this reason, the standard import/export
functions cannot be used.  Instead, a direct copy approach is taken.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@829 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: fe123f6813d552b86ea1f94a6ada96316114a674