Implemented central-difference scheme for computing the Jacobian.
There is now a virtual function in the PeridigmNS::Material base class
called computeJacobian.  The idea is that this function can be
implemented in a material model that derives from this base class,
allowing for the use of an analytic Jacobian.

If computeJacobian is not implemented in a derived class, then by
default computeJacobian calls computeFiniteDifferenceJacobian.  The
function computeFiniteDifferenceJacobian takes an argument
FiniteDifferenceScheme, which can be either FORWARD_DIFFERENCE or
CENTRAL_DIFFERENCE.  Currently it is hard-coded to CENTRAL_DIFFERENCE,
but Peridigm could be set up to read the finite-difference scheme
from the input deck.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@832 d9297710-12f9-482b-b11f-fb677dc88bbb