Added regression test that exercises new functionality for reading Exodus mesh files.
This test is a copy of the simple compression quasi-statics
regression test.  The difference is that, instead of using
PdQuickGrid to generate the discretization, the discretization
is created from an Exodus file (hex mesh) that was created
in Cubit.

Note that the Cubit and PdQuickGrid assign node/element numbers
differently.  Thus the output files cannot be compared directly
using vtkDiff.py.  They can be visually compared in ParaView,
however, or manually compared by opening the .vtu files in an
editor.

The new test currently runs only in serial.  The parallel
version is forthcoming.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@835 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: dfd7d48d690fda49b8e9e29e608c2bcc64b774e2